[GLUTEN-5484] add missing tests for clickhouse
